+++ src/main/java/tesis/entities/dtos/home/ItemHomeCarrouselBottom.java
@@ -1,0 +1,6 @@
+package tesis.entities.dtos.home;
+
+import tesis.entities.dtos.home.ItemHome;
+
+public class ItemHomeCarrouselBottom extends ItemHome {
+}
