+++ src/main/java/tesis/entities/dtos/home/ItemHomeHurryUp.java
@@ -1,6 +1,7 @@
-package tesis.entities.dtos.item;
+package tesis.entities.dtos.home;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import tesis.entities.dtos.home.ItemHome;
 
 public class ItemHomeHurryUp extends ItemHome {
