--- src/main/java/tesis/entities/dtos/item/ItemHomeCarrouselBottom.java
@@ -1,4 +1,0 @@
-package tesis.entities.dtos.item;
-
-public class ItemHomeCarrouselBottom extends ItemHome {
-}
